+++ src/DiscordMusic.Core/Discord/Cache/MusicCache.cs
@@ -99,6 +99,23 @@ internal class MusicCache(
     private async Task<ErrorOr<Success>> IndexAsync(CancellationToken ct)
     {
         var cacheLocation = GetCacheLocation(logger, cacheOptions, fileSystem);
+
+        try
+        {
+            if (!cacheLocation.Exists())
+            {
+                logger.LogDebug(
+                    "Cache location {Location} does not exist. Creating directory.",
+                    cacheLocation
+                );
+                cacheLocation.Create();
+            }
+        }
+        catch (Exception e)
+        {
+            return Error.Unexpected(e.Message);
+        }
+
         return await _fileCache.IndexAsync(cacheLocation, ct);
     }
 
@@ -108,35 +125,56 @@ internal class MusicCache(
         IFileSystem fileSystem
     )
     {
-        if (string.IsNullOrWhiteSpace(cacheOptions.Value.Location))
+        if (!string.IsNullOrWhiteSpace(cacheOptions.Value.Location))
         {
-            var appData = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
-            var cacheDir = fileSystem.DirectoryInfo.New(
-                Path.Combine(appData, "discord-music", "cache")
+            logger.LogDebug(
+                "Using {Location} from environment variable or config file as cache location",
+                cacheOptions.Value.Location
             );
-            logger.LogWarning("Cache location not set, using default {Location}", cacheDir);
-
-            if (cacheDir.Exists())
-            {
-                return cacheDir;
-            }
-
-            logger.LogDebug("Creating cache directory {Location}", cacheDir);
-            cacheDir.Create();
-
-            return cacheDir;
-        }
-
-        var cacheLocation = fileSystem.DirectoryInfo.New(cacheOptions.Value.Location);
-
-        if (cacheLocation.Exists())
-        {
+            var cacheLocation = fileSystem.DirectoryInfo.New(cacheOptions.Value.Location);
+            logger.LogDebug("Final env or config location {Location}", cacheLocation);
             return cacheLocation;
         }
 
-        logger.LogDebug("Creating cache directory {Location}", cacheLocation);
-        cacheLocation.Create();
+        var xdgCacheHome = Environment.GetEnvironmentVariable("XDG_CACHE_HOME");
+        if (!string.IsNullOrWhiteSpace(xdgCacheHome))
+        {
+            logger.LogDebug(
+                "Using XDG_CACHE_HOME '{XDG_CACHE_HOME}' as cache location",
+                xdgCacheHome
+            );
+            var cacheDir = fileSystem.DirectoryInfo.New(
+                Path.Combine(xdgCacheHome, "bycrookie", "discord-music")
+            );
+            logger.LogDebug("Final XDG cache location {Location}", cacheDir);
+            return cacheDir;
+        }
 
-        return cacheLocation;
+        if (OperatingSystem.IsWindows())
+        {
+            logger.LogDebug("Using windows local app data location as cache location");
+            var windows = fileSystem.DirectoryInfo.New(
+                Path.Combine(
+                    Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
+                    "bycrookie",
+                    "discord-music",
+                    "cache"
+                )
+            );
+            logger.LogDebug("Final windows location {Location}", windows);
+            return windows;
+        }
+
+        logger.LogDebug("Using unix home as cache location");
+        var unix = fileSystem.DirectoryInfo.New(
+            Path.Combine(
+                Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
+                ".cache",
+                "bycrookie",
+                "discord-music"
+            )
+        );
+        logger.LogDebug("Final unix location {Location}", unix);
+        return unix;
     }
 }