+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 **Changes:**
 
-- update to dotnet 10
+- BREAKING: change config/cache sources (no changes required when using containers)
+- add [xdg](https://specifications.freedesktop.org/basedir/latest/) support